+++ lib/src/constants/constants.dart
@@ -3,6 +3,8 @@ import 'package:path/path.dart';
 const defaultClassName = 'S';
 const defaultMainLocale = 'en';
 final defaultArbDir = join('lib', 'l10n');
+final defaultOutputDir = join('lib', 'generated');
+const defaultUseDeferredLoading = false;
 const defaultUploadOverwrite = false;
 const defaultUploadAsReviewed = false;
 const defaultDownloadEmptyAs = 'empty';