+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to the "flutter-intl" extension will be documented in this f
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.9.0 - 2020-10-19
+
+- Make generated directory path configurable
+
+- Extend configuration with deferred loading parameter
+
+- Ignore common lint warnings for the `l10n.dart` file
+
 ## 1.8.0 - 2020-10-09
 
 - Extend Localizely configuration with the download_empty_as parameter used for setting a fallback for empty translations during download